Refactor deepagents assembly and simplify execution graph
