+++ backend/app/services/deepagents/assembly/__init__.py
@@ -1,0 +1,31 @@
+from app.services.deepagents.assembly.backend_factory import _backend_root, build_agent_backend_factory
+from app.services.deepagents.assembly.output_mapping import (
+    DeepAgentsCancelled,
+    DeepAgentsLoopResult,
+    DeepAgentsToolRun,
+    build_loop_result,
+    map_tool_runs,
+    parse_capabilities_file,
+)
+from app.services.deepagents.assembly.prompt import build_system_prompt, tool_description
+from app.services.deepagents.assembly.skill_mounts import (
+    SkillMountSnapshot,
+    get_skill_mount_snapshot,
+)
+from app.services.deepagents.assembly.tool_registry import build_chat_tools
+
+__all__ = [
+    "DeepAgentsCancelled",
+    "DeepAgentsLoopResult",
+    "DeepAgentsToolRun",
+    "SkillMountSnapshot",
+    "_backend_root",
+    "build_agent_backend_factory",
+    "build_chat_tools",
+    "build_loop_result",
+    "build_system_prompt",
+    "get_skill_mount_snapshot",
+    "map_tool_runs",
+    "parse_capabilities_file",
+    "tool_description",
+]
